facereclib/__init__.py
#!/usr/bin/env python
# vim: set fileencoding=utf-8 :
# Manuel Guenther <Manuel.Guenther@idiap.ch>

"""Library for face recognition experiments"""

from . import databases
from . import preprocessing
from . import features
from . import tools
from . import utils
from . import toolchain

from . import script

from . import tests

def get_config():
  """Returns a string containing the configuration information.
  """
  import bob.extension
  return bob.extension.get_config(__name__)


# gets sphinx autodoc done right - don't remove it
__all__ = [_ for _ in dir() if not _.startswith('_')]
